// src/enums.ts
export enum Color {
    Black = "rgb(0, 0, 0)",
    Green = "rgb(0, 215, 0)"
}

export enum KeyCode {
    Space = 'Space',
    ArrowUp = 'ArrowUp',
    ArrowDown = 'ArrowDown',
    ArrowLeft = 'ArrowLeft',
    ArrowRight = 'ArrowRight'
}

export enum GameObjectType {
    Snake = "snake",
    Apple = "apple"
}